+++ CHANGELOG.md
@@ -15,8 +15,10 @@ The shortcode for this plugin is [IQBible].
 - uninstall.php
 ### Changed
 - reloadChapterContent to loadChapterContent for clarity as it is not always 'reloading' (e.g., initial page load)
+- - credits.txt (is properly in readme.txt for WP Mktplc)
 ### Fixed
 - Fixed clearbooksession() in version update to return a promise before executing loadChapterContent (to avoid race condition)
+- Proper $wpdb->prepare() in cache clearing forms (2)
 
 ## [1.0.0-alpha-7] - 2025-04-26
 ### Added
@@ -34,7 +36,6 @@ The shortcode for this plugin is [IQBible].
 - Footer (added dialogs to shortcodes) as data is not suited for the WP site owner's visitors. Instead, display version data in admin-settings.php
 - Renamed generic start_session to iqbible_start_session to avoid potential conflicts
 - enqueue_dashicons to iqbible_enqueue_dashicons
-- credits.txt (is properly in readme.txt for WP Mktplc)
 
 ## [1.0.0-alpha-6] - 2025-04-24 - 2025-04-25
 ### Added